User commit

+++ src/main/java/com/example/demo/service/impl/ValidationServiceImpl.java
@@ -3,6 +3,7 @@ import java.util.*;
 import com.example.demo.service.ValidationService;
 import com.example.demo.repository.ValidationRepo;
 import com.example.demo.entity.ValidationEntity;
+import com.example.demo.exception.ValidationException;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -15,7 +16,7 @@ public class ValidationServiceImpl implements ValidationService{
                return dent.save(stu);
           }
           @Override
-          public ValidationEntity finddata(int id){
+          public ValidationEntity finddata(Long id){
                return dent.findById(id).orElseThrow(()->new ValidationException("Invalid Id"+id));
           }
 }
